Corrected inheritance
